rename: User 예외처리 수정

+++ server/src/main/java/com/codestates/server/user/service/UserService.java
@@ -2,7 +2,9 @@ package com.codestates.server.user.service;
 
 import com.codestates.server.answer.entity.Answer;
 import com.codestates.server.answer.repository.AnswerRepository;
-import com.codestates.server.auth.utils.AuthUserUtils;
+import com.codestates.server.exception.BusinessLogicException;
+import com.codestates.server.exception.ExceptionCode;
+import com.codestates.server.security.auth.utils.AuthUserUtils;
 import com.codestates.server.question.entity.Question;
 import com.codestates.server.question.repository.QuestionRepository;
 import com.codestates.server.security.auth.utils.CustomAuthorityUtils;
@@ -61,7 +63,7 @@ public class UserService {
 
         // 로그인 User의 아이디와 회원정보를 가진 user의 아이디가 다르면 예외 던지기
         if(!getLoginUser().getUserId().equals(getUser.getUserId()))
-            throw new RuntimeException();   // 🚨 예외처리
+            throw new BusinessLogicException(ExceptionCode.UNAUTHORIZED_USER);   // 🚨 예외처리
 
         Optional.ofNullable(user.getUserName())
                 .ifPresent(name -> getUser.setUserName(user.getUserName()));
@@ -130,6 +132,10 @@ public class UserService {
     public void deleteUser(Long userId) {
         User getUser = getVerifiedUser(userId);
 
+        // 로그인 User의 아이디와 회원정보를 가진 user의 아이디가 다르면 예외 던지기
+        if(!getLoginUser().getUserId().equals(getUser.getUserId()))
+            throw new BusinessLogicException(ExceptionCode.UNAUTHORIZED_USER);   // 🚨 예외처리
+
         userRepository.delete(getUser);
     }
 
@@ -140,7 +146,7 @@ public class UserService {
         Optional<User> user = userRepository.findById(userId);
 
         User getUser =
-                user.orElseThrow(() -> new RuntimeException());
+                user.orElseThrow(() -> new BusinessLogicException(ExceptionCode.USER_NOT_FOUND));
         // 🚨 예외 처리
         return getUser;
     }
@@ -151,14 +157,13 @@ public class UserService {
         Optional<User> user = userRepository.findByEmail(email);
 
         if(user.isPresent())
-            throw new RuntimeException();
+            throw new BusinessLogicException(ExceptionCode.USER_EXISTS);
         // 🚨 예외 처리
     }
 
     // 로그인한 User를 가지고 오는 메서드
     public User getLoginUser() {
         return userRepository.findByEmail(AuthUserUtils.getAuthUser().getName())
-
-                .orElseThrow(() -> new RuntimeException()); // 🚨 예외처리
+                .orElseThrow(() -> new BusinessLogicException(ExceptionCode.USER_NOT_FOUND)); // 🚨 예외처리
     }
 }
